+++ lib/pages/inicio.dart
@@ -93,7 +93,7 @@ class Inicio extends StatelessWidget {
             const SizedBox(height: 20),
             TextButton(
               onPressed: () {
-                Navigator.pop(context, '/home');
+                Navigator.pushNamed(context, '/home');
               },
               child: const Text("SKIP IT FOR NOW", style: TextStyle(color: Colors.grey, fontWeight: FontWeight.w500)),
             ),
