+++ lib/widgets/home_sections_card.dart
@@ -34,15 +34,15 @@ class _MindSectionHeader extends StatelessWidget {
           style: TextStyle(
             fontSize: 20,
             fontWeight: FontWeight.bold,
-            color: Colors.grey[800],
+            color: Colors.black,
           ),
         ),
         SizedBox(height: 8),
         Text(
           "Choose what fits you best!",
           style: TextStyle(
-            fontSize: 14,
-            color: Colors.grey[600],
+            fontSize: 16,
+            color: Colors.grey[900],
           ),
         ),
       ],
@@ -67,26 +67,26 @@ class _MindCardsGrid extends StatelessWidget {
       children: [
         MindCard(
           title: 'Weekly Challenge',
-          color: Colors.blue[400]!,
-          icon: Icons.calendar_today,
+          color: Color(0xFF6389E2),
+          url: '/weeklyChallenge',
           onTap: () => onCardTap(CardType.weeklyChallenge),
         ),
         MindCard(
           title: 'Personality Quiz',
-          color: Colors.pink[400]!,
-          icon: Icons.quiz,
+          color: Color(0xFFED6275),
+          url: '/personalityQuiz',
           onTap: () => onCardTap(CardType.personalityQuiz),
         ),
         MindCard(
           title: 'Wish me Luck',
-          color: Colors.pink[400]!,
-          icon: Icons.favorite,
+          color: Color(0xFFED6275),
+          url: '/whishMeLuck',
           onTap: () => onCardTap(CardType.wishMeLuck),
         ),
         MindCard(
           title: 'Map',
-          color: Colors.blue[400]!,
-          icon: Icons.map,
+          color: Color(0xFF6389E2),
+          url: '/map',
           onTap: () => onCardTap(CardType.map),
         ),
       ],
@@ -97,28 +97,30 @@ class _MindCardsGrid extends StatelessWidget {
 class MindCard extends StatelessWidget {
   final String title;
   final Color color;
-  final IconData icon;
+  final String url;
   final VoidCallback onTap;
 
   const MindCard({
     Key? key,
     required this.title,
     required this.color,
-    required this.icon,
     required this.onTap,
+    required this.url,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: (){
+        Navigator.pushNamed(context, url);
+      },
       child: Container(
         decoration: BoxDecoration(
           color: color,
           borderRadius: BorderRadius.circular(20),
           boxShadow: [
             BoxShadow(
-              color: color.withOpacity(0.3),
+              color: color.withValues(alpha: 0.3),
               spreadRadius: 1,
               blurRadius: 8,
               offset: Offset(0, 4),
@@ -129,14 +131,8 @@ class MindCard extends StatelessWidget {
           padding: EdgeInsets.all(15),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.center,
             children: [
-              Icon(
-                icon,
-                color: Colors.white,
-                size: 24,
-              ),
-              SizedBox(height: 8),
               Text(
                 title,
                 style: TextStyle(
